divided JS code into modules closes #20
